# [SR-610][stdlib] Use SequenceType._initializeTo where possible