--- a/stdlib/public/core/Sequence.swift
+++ b/stdlib/public/core/Sequence.swift
@@ -194,10 +194,14 @@ public protocol SequenceType {
   /// in the same order.
   func _copyToNativeArrayBuffer() -> _ContiguousArrayBuffer<Generator.Element>
 
-  /// Copy a Sequence into an array, returning one past the last
-  /// element initialized.
-  func _initializeTo(ptr: UnsafeMutablePointer<Generator.Element>)
-    -> UnsafeMutablePointer<Generator.Element>
+  /// Copy the contents of `self` into uninitialized contiguous memory
+  /// of size `capacity`, returning one past the last element
+  /// initialized, or a `nil` pointer if the copy cannot be safely
+  /// made given `capacity`.
+  @warn_unused_result
+  func _initializeTo(
+    start: UnsafeMutablePointer<Generator.Element>, capacity: Int
+  ) -> UnsafeMutablePointer<Generator.Element>
 }
 
 /// A default generate() function for `GeneratorType` instances that
@@ -604,14 +608,11 @@ public func underestimateCount<T : SequenceType>(x: T) -> Int {
 }
 
 extension SequenceType {
-  public func _initializeTo(ptr: UnsafeMutablePointer<Generator.Element>)
-    -> UnsafeMutablePointer<Generator.Element> {
-    var p = UnsafeMutablePointer<Generator.Element>(ptr)
-    for x in GeneratorSequence(self.generate()) {
-      p.initialize(x)
-      p += 1
-    }
-    return p
+  @warn_unused_result
+  public func _initializeTo(
+    start: UnsafeMutablePointer<Generator.Element>, capacity: Int
+  ) -> UnsafeMutablePointer<Generator.Element> {
+    return nil
   }
 }