--- a/stdlib/public/core/UnsafePointer.swift.gyb
+++ b/stdlib/public/core/UnsafePointer.swift.gyb
@@ -274,7 +274,21 @@ ${comment}
   public func initializeFrom<
     C : CollectionType where C.Generator.Element == Memory
   >(source: C) {
-    source._initializeTo(self)
+    let count: Int = numericCast(source.count)
+    let end = source._initializeTo(self, capacity: count)
+    if end != nil {
+      _debugPrecondition(end == self + count)
+    }
+    else {
+      var p = self
+      var i = source.startIndex
+      for _ in 0..<count {
+        p.initialize(source[i])
+        i = i.successor()
+        p += 1
+      }
+      _expectEnd(i, source)
+    }
   }
 
   /// Assign from `count` values beginning at `source` into initialized