(This is an updated version of PR #1084.)

Change the `SequenceType._initializeTo` entry point to support memory safety better,
and use it in more places where `Array` elements are initialized in bulk:
- Conversion of sequences and collections to `Array`
- `<ArrayType> += <SequenceType>`
- `ArrayType.replaceRange` (when used for insertion)

The speedup achieved is highly sensitive to the mood of the optimizer. The most remarkable result is that insertion of an `Array` into another `Array` can be 4-35x faster depending on the element type. Besides `Array`, `ArraySlice` and `ContiguousArray` also have custom `_initializeTo` implementations, but interestingly, only `Array` seems to be meaningfully impacted by these changes. (The behavior of the
optimizer can be highly unintuitive, making attempts at micro-optimization less useful; see SR-653.)

When `_initializeTo` isn't available, prefer `Generator` to looping over index ranges when iterating over elements of a collection. This speeds up the above operations for collections with complex indexes. 
For example, operations on standard hashed collections typically improve by about 10-40%.