@lorentey 2016-02-03
(This is an updated version of PR #1084.)