@lorentey 2016-02-03
I attached detailed benchmarking results as [SR-610.xlsx](https://github.com/apple/swift/files/116531/SR-610.xlsx). The benchmark measures execution time of the following operations, for each of `Array<Int>`, `Array<String>`, `Array<(Int, String)>`, `ArraySlice<Int>`, `ArraySlice<(Int, String)>`, `Dictionary<Int, String>`, `Set<Int>`, `Set<String>`, `Deque<Int>`, `Deque<(Int, String)>`, and `Map<Int, String>`:

1. Converting the collection to `Array` via `Array(thing)`
2. Converting the collection to `Array` using `Array(GeneratorSequence(thing.generate()))`
3. Appending the contents of the collection to the end of an `Array` using `+=`
4. Inserting the contents of the collection in the middle of an `Array` using `insertContentsOf`

The `Deque` type is a custom double-ended queue built on a circular buffer. `Map` is an associative collection like `Dictionary`, but it uses `Comparable` keys and it is built on an in-memory b-tree. As a sanity check, the results also contain control measurements marked "ZZZ Control" that do not use modified codepaths, and should measure the same before/after the patch is applied.

The source code for my benchmark utility is at
[SR610Benchmark.zip](https://github.com/apple/swift/files/116543/SR610Benchmark.zip).