@gribozavr 2016-02-03
@swift-ci Please test