@jckarter 2016-02-04
@aschwaighofer and/or @nadavrot will probably want to look at this. I know we've gone back and forth on some of these micro optimizations as the high level optimizations passes have improved.