@nadavrot 2016-02-04
@lorentey It looks like your commit squashed a bunch of different changes (to initializeTo,  to stdlibunittests, etc). Can you split the commit into multiple patches that can be reviewed independently?