@lorentey 2016-02-04
@nadavrot Absolutely! Would these three phases be easier to review?

- PR 1: Update `_initializeTo` to the new signature everywhere.
- PR 2: Use the new `_initializeTo` in more operations.
- PR 3: Prefer `generate()` to indexing.