@lorentey 2016-02-04
@nadavrot I split the PR into three distinct commits. I hope it will be easier to review now!