@lorentey 2016-02-04
@swift-ci Please test