@gribozavr 2016-02-15
@lorentey I'm sorry for not responding!  Unfortunately I was busy.  I will try to review this PR.

One thing that you could do is to run the performance testsuite that was recently added to the repository (built binary at `bin/Benchmark_O`).  You'd need to build with `-R --no-swift-stdlib-assertions` to get meaningful benchmark results.