@lorentey 2016-02-16
No worries! This isn't urgent at all, I just don't want to forget about it.

I ran `utils/build-script -R --no-swift-stdlib-assertions --benchmark` a couple of times, collecting the `-O` results in a spreadsheet ([PDF](https://github.com/apple/swift/files/131832/Comparison.pdf), [Excel](https://github.com/apple/swift/files/131833/Comparison.xlsx)).

The proposed change seems to be a slight overall improvement (1.5% total) in optimized builds. 17 out of  78 tests show larger than 3% improvement, while 4 others (`ClassArrayGetter`, `RangeAssignment`, `RGBHistogram` and `Sim2DArray`) have at least the same delta in the negative direction. Of the latter, `RGBHistogram` is particularly interesting, because for some reason it exhibits a whopping 35% slowdown; the three others are slower by a relatively boring 6-8%.

I haven't analyzed `-Onone` results yet, but anectodally I did notice major (30-100%) slowdowns for `MapReduce`, `NopDeinit`, and `RangeAssignment`. (The rest of the tests seem to be a mixed bag when optimization is off.)

I will look into the tests to see if I can explain these outliers and hopefully fix them.