@gribozavr 2016-02-16
It would be good if you could look into RGBHistogram.  The first thing to do is to increase the number of samples to ensure that the difference is not caused by noise: 

```
./bin/Benchmark_O --num-samples=10 RGBHistogram
```