@nadavrot 2016-02-16
I don’t think that it is very likely that a change to SequenceType improved the performance of the Calculator test by so much. I recommend running the benchmark again with a higher number of iterations. 

> On Feb 15, 2016, at 7:47 PM, Károly Lőrentey <notifications@github.com> wrote:
> 
> No worries! This isn't urgent at all, I just don't want to forget about it.
> 
> I ran utils/build-script -R --no-swift-stdlib-assertions --benchmark a couple of times, collecting the -O results in a spreadsheet (PDF <https://github.com/apple/swift/files/131832/Comparison.pdf>, Excel <https://github.com/apple/swift/files/131833/Comparison.xlsx>).
> 
> The proposed change seems to be a slight overall improvement (1.5% total) in optimized builds. 17 out of 78 tests show larger than 3% improvement, while 4 others (ClassArrayGetter, RangeAssignment, RGBHistogram and Sim2DArray) have at least the same delta in the negative direction. Of the latter, RGBHistogram is particularly interesting, because for some reason it exhibits a whopping 35% slowdown; the three others are slower by a relatively boring 6-8%.
> 
> I haven't analyzed -Onone results yet, but anectodally I did notice major (30-100%) slowdowns for MapReduce, NopDeinit, and RangeAssignment. (The rest of the tests seem to be a mixed bag when optimization is off.)
> 
> I will look into the tests to see if I can explain these outliers and hopefully fix them.
> 
> —
> Reply to this email directly or view it on GitHub <https://github.com/apple/swift/pull/1194#issuecomment-184504629>.
>