@lorentey 2016-02-16
@gribozavr @nadavrot Increasing the number of samples & iterations eliminates the slowdown for `ClassArrayGetter`, but keeps it for the rest of the big losers, including `RGBHistogram`.

The `Calculator` result is reliably repeatable; this is perplexing. I'll add markers to code paths I touched to see if they're somehow involved. 

`RGBHistogram` calls `sort` on a dictionary of integers, which seems like a nice candidate for me to start looking for regressions.

Raw results for 100 samples and 100 iterations are below:

#### Before

`./build/master/swift-macosx-x86_64/bin/Benchmark_O --num-samples=100 --num-iters=100 RGBHistogram Calculator ClassArrayGetter RangeAssignment Sim2DArray`

#|TEST|SAMPLES|MIN(us)|MAX(us)|MEAN(us)|SD(us)|MEDIAN(us)
--- | --- | --- | --- | --- | --- | --- | ---
18|Calculator|100|44|85|46|6|45
21|ClassArrayGetter|100|25|41|27|3|26
56|RGBHistogram|100|3147|3411|3251|60|3246
57|RangeAssignment|100|332|471|351|20|347
64|Sim2DArray|100|777|902|809|29|802
Totals|5||4325|4910|4484|0|0

#### After

`./build/sr610/swift-macosx-x86_64/bin/Benchmark_O --num-samples=100 --num-iters=100 RGBHistogram Calculator ClassArrayGetter RangeAssignment Sim2DArray`

#|TEST|SAMPLES|MIN(us)|MAX(us)|MEAN(us)|SD(us)|MEDIAN(us)
--- | --- | --- | --- | --- | --- | --- | ---
18|Calculator|100|38|71|41|5|40
21|ClassArrayGetter|100|25|40|26|2|25
56|RGBHistogram|100|4310|4737|4442|82|4427
57|RangeAssignment|100|352|427|371|17|368
64|Sim2DArray|100|786|964|830|40|828
Totals|5||5511|6239|5710|0|0