@lorentey 2016-02-16
Update on `Calculator`: taking the `master` branch and adding a `print()` call to `CollectionType.map` in `Collections.swift` is enough to reproduce the speedup. This is true even though the `map` function is not actually called by the benchmark case. Hm.

This is with commit 2c0fd720 from yesterday; I'm going to pull the latest `master` and retry.