@lorentey 2016-02-17
I looked into `RGBHistogram`. The slowdown is due to the array conversion in `Dictionary.sort`, and can be entirely attributed to a single inlining decision: when I add `@inline(__always)` on `DictionaryGenerator.next()`, the slowdown disappears.

This suggests the same inlining decision are also responsible for the 10-40% *speedup* I measured in my own `Dictionary` benchmarks, which were all using native dictionaries.

I still expect the new code to behave meaningfully better for bridged `NSDictionary` instances, where the generator is able to use fast enumeration, while creating indexes is slow.